ID token expiry time based on local clock
The OIDC spec retained the OAuth 2 `expires_in` response parameter which
relates to the issued `access_token`. Unfortunately the spec did not
introduce an equivalent parameter for `id_token`, which is presumably
why the current code just uses the token `exp` value. This is unfortunate
and feels like a deficiency for the OIDC spec, since it makes it harder
to accommodate for local clock drift.

This change works out how many seconds the ID token expires in by
subtracting the issued at time from the expires at time. We can then add
this to `Date.now()` to calculate an `expiresAt` time which is based on
the local clock. Note that this is already the case for the Access Token,
which adds the `expires_in` response value to `Date.now()`.

Also note that this is only appropriate if the OIDC endpoint mints ID
tokens each request (i.e. there is no caching involved, etc).